# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Car Idle Game API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface LeaderboardEntry {
  id: number;
  playerId: string;
  playerName: string;
  totalMiles: number;
  prestigeLevel: number;
  carsOwned: number;
  updatedAt: string;
}

export interface ScoreInput {
  playerId: string;
  playerName: string;
  totalMiles: number;
  prestigeLevel: number;
  carsOwned: number;
}

export interface PlayerRank {
  rank: number;
  entry: LeaderboardEntry;
}

export interface GameSave {
  id: number;
  playerId: string;
  saveData: string;
  updatedAt: string;
}

export interface GameSaveInput {
  saveData: string;
}

export interface Trade {
  id: number;
  offerPlayerId: string;
  offerPlayerName: string;
  offeredCarId: string;
  offeredCarName: string;
  wantedCarId: string;
  wantedCarName: string;
  /** @nullable */
  acceptPlayerId?: string | null;
  status: string;
  createdAt: string;
}

export interface TradeInput {
  offerPlayerId: string;
  offerPlayerName: string;
  offeredCarId: string;
  offeredCarName: string;
  wantedCarId: string;
  wantedCarName: string;
}

export interface TradeAcceptInput {
  acceptPlayerId: string;
}

export type GetLeaderboardParams = {
limit?: number;
};

export type ListTradesParams = {
/**
 * Filter by offering player
 */
playerId?: string;
};