# lib/api-zod/src/generated/types/gameSave.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Car Idle Game API
 * OpenAPI spec version: 0.1.0
 */

export interface GameSave {
  id: number;
  playerId: string;
  saveData: string;
  updatedAt: string;
}