# lib/api-zod/src/generated/types/playerRank.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Car Idle Game API
 * OpenAPI spec version: 0.1.0
 */
import type { LeaderboardEntry } from './leaderboardEntry';

export interface PlayerRank {
  rank: number;
  entry: LeaderboardEntry;
}