# lib/api-zod/src/generated/types/scoreInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Car Idle Game API
 * OpenAPI spec version: 0.1.0
 */

export interface ScoreInput {
  playerId: string;
  playerName: string;
  totalMiles: number;
  prestigeLevel: number;
  carsOwned: number;
}