# lib/api-zod/src/generated/types/trade.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Car Idle Game API
 * OpenAPI spec version: 0.1.0
 */

export interface Trade {
  id: number;
  offerPlayerId: string;
  offerPlayerName: string;
  offeredCarId: string;
  offeredCarName: string;
  wantedCarId: string;
  wantedCarName: string;
  /** @nullable */
  acceptPlayerId?: string | null;
  status: string;
  createdAt: string;
}