# lib/api-zod/src/generated/types/tradeInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Car Idle Game API
 * OpenAPI spec version: 0.1.0
 */

export interface TradeInput {
  offerPlayerId: string;
  offerPlayerName: string;
  offeredCarId: string;
  offeredCarName: string;
  wantedCarId: string;
  wantedCarName: string;
}